moved addmodal to seperate file
